Add a generic JSON consumer
Requires kafka-python, curses, and a few fairly generic classes.

Takes one argument on the command line; the name of the kafka topic to take input from. (If this is not provided, it will try to read from a kafka topic called 'default', which is harmless but probably not what you're looking for). To exit, give it a keypress - any keypress.

Graphs are not guaranteed to be meaningful, However, it should help to test almost any producer.